+++ tests/create3.sh
@@ -55,7 +55,7 @@ _cj_internal_conf()
 test_cj_conf_001()
 {
 	# level 0
-	_cj_conf new-pot 11.1 inherit NO 0 inherit multi
+	_cj_conf new-pot 11.1 inherit inherit 0 inherit multi
 	assertEquals "return code" "0" "$?"
 	assertEquals "fscomp args1" "zpot/bases/11.1 /tmp/jails/new-pot/m" "$(sed '1!d' /tmp/jails/new-pot/conf/fscomp.conf)"
 	assertEquals "fscomp args2" "zpot/bases/11.1/usr.local /tmp/jails/new-pot/m/usr/local" "$(sed '2!d' /tmp/jails/new-pot/conf/fscomp.conf)"
@@ -64,7 +64,8 @@ test_cj_conf_001()
 	assertEquals "pot.base" "pot.base=11.1" "$(grep ^pot.base /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "osrelease" "osrelease=\"11.1-RELEASE\"" "$(grep ^osrelease /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "pot.potbase" "pot.potbase=" "$(grep ^pot.potbase /tmp/jails/new-pot/conf/pot.conf)"
-	assertEquals "ip4" "ip4=inherit" "$(grep ^ip4= /tmp/jails/new-pot/conf/pot.conf)"
+	assertEquals "network_type" "network_type=inherit" "$(grep ^network_type= /tmp/jails/new-pot/conf/pot.conf)"
+	assertEquals "ip" "" "$(grep ^ip= /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "vnet" "vnet=false" "$(grep ^vnet= /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "mkdir calls" "1" "$MKDIR_CALLS"
 	assertEquals "mkdir arg2" "${POT_FS_ROOT}/jails/new-pot/conf" "$MKDIR_CALL1_ARG2"
@@ -78,7 +79,7 @@ test_cj_conf_001()
 
 test_cj_conf_002()
 {
-	_cj_conf new-pot 11.1 inherit NO 1 inherit multi
+	_cj_conf new-pot 11.1 inherit inherit 1 inherit multi
 	assertEquals "return code" "0" "$?"
 	assertEquals "fscomp args1" "zpot/bases/11.1 /tmp/jails/new-pot/m ro" "$(sed '1!d' /tmp/jails/new-pot/conf/fscomp.conf)"
 	assertEquals "fscomp args2" "zpot/jails/new-pot/usr.local /tmp/jails/new-pot/m/usr/local zfs-remount" "$(sed '2!d' /tmp/jails/new-pot/conf/fscomp.conf)"
@@ -87,7 +88,8 @@ test_cj_conf_002()
 	assertEquals "pot.base" "pot.base=11.1" "$(grep ^pot.base /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "osrelease" "osrelease=\"11.1-RELEASE\"" "$(grep ^osrelease /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "pot.potbase" "pot.potbase=" "$(grep ^pot.potbase /tmp/jails/new-pot/conf/pot.conf)"
-	assertEquals "ip4" "ip4=inherit" "$(grep ^ip4= /tmp/jails/new-pot/conf/pot.conf)"
+	assertEquals "network_type" "network_type=inherit" "$(grep ^network_type= /tmp/jails/new-pot/conf/pot.conf)"
+	assertEquals "ip" "" "$(grep ^ip= /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "vnet" "vnet=false" "$(grep ^vnet= /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "mkdir calls" "1" "$MKDIR_CALLS"
 	assertEquals "mkdir arg2" "${POT_FS_ROOT}/jails/new-pot/conf" "$MKDIR_CALL1_ARG2"
@@ -105,7 +107,7 @@ test_cj_conf_003()
 	echo "zpot/bases/11.1 /tmp/jails/test-pot/m ro" >> /tmp/jails/test-pot/conf/fscomp.conf
 	echo "zpot/jails/test-pot/usr.local /tmp/jails/test-pot/m/usr/local zfs-remount" >> /tmp/jails/test-pot/conf/fscomp.conf
 	echo "zpot/jails/test-pot/custom /tmp/jails/test-pot/m/opt/custom zfs-remount" >> /tmp/jails/test-pot/conf/fscomp.conf
-	_cj_conf new-pot 11.1 inherit NO 1 inherit multi test-pot
+	_cj_conf new-pot 11.1 inherit inherit 1 inherit multi test-pot
 	assertEquals "return code" "0" "$?"
 	assertEquals "fscomp args1" "zpot/bases/11.1 /tmp/jails/new-pot/m ro" "$(sed '1!d' /tmp/jails/new-pot/conf/fscomp.conf)"
 	assertEquals "fscomp args2" "zpot/jails/new-pot/usr.local /tmp/jails/new-pot/m/usr/local zfs-remount" "$(sed '2!d' /tmp/jails/new-pot/conf/fscomp.conf)"
@@ -115,7 +117,8 @@ test_cj_conf_003()
 	assertEquals "pot.base" "pot.base=11.1" "$(grep ^pot.base /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "osrelease" "osrelease=\"11.1-RELEASE\"" "$(grep ^osrelease /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "pot.potbase" "pot.potbase=test-pot" "$(grep ^pot.potbase /tmp/jails/new-pot/conf/pot.conf)"
-	assertEquals "ip4" "ip4=inherit" "$(grep ^ip4= /tmp/jails/new-pot/conf/pot.conf)"
+	assertEquals "network_type" "network_type=inherit" "$(grep ^network_type= /tmp/jails/new-pot/conf/pot.conf)"
+	assertEquals "ip" "" "$(grep ^ip= /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "vnet" "vnet=false" "$(grep ^vnet= /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "mkdir calls" "1" "$MKDIR_CALLS"
 	assertEquals "mkdir arg2" "${POT_FS_ROOT}/jails/new-pot/conf" "$MKDIR_CALL1_ARG2"
@@ -133,7 +136,7 @@ test_cj_conf_004()
 	echo "zpot/bases/11.1 /tmp/jails/test-pot/m ro" >> /tmp/jails/test-pot/conf/fscomp.conf
 	echo "zpot/jails/test-pot/usr.local /tmp/jails/test-pot/m/usr/local zfs-remount" >> /tmp/jails/test-pot/conf/fscomp.conf
 	echo "zpot/jails/test-pot/custom /tmp/jails/test-pot/m/opt/custom zfs-remount" >> /tmp/jails/test-pot/conf/fscomp.conf
-	_cj_conf new-pot 11.1 inherit NO 2 inherit multi test-pot
+	_cj_conf new-pot 11.1 inherit inherit 2 inherit multi test-pot
 	assertEquals "return code" "0" "$?"
 	assertEquals "fscomp args1" "zpot/bases/11.1 /tmp/jails/new-pot/m ro" "$(sed '1!d' /tmp/jails/new-pot/conf/fscomp.conf)"
 	assertEquals "fscomp args2" "zpot/jails/test-pot/usr.local /tmp/jails/new-pot/m/usr/local ro" "$(sed '2!d' /tmp/jails/new-pot/conf/fscomp.conf)"
@@ -143,7 +146,7 @@ test_cj_conf_004()
 	assertEquals "pot.base" "pot.base=11.1" "$(grep ^pot.base /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "osrelease" "osrelease=\"11.1-RELEASE\"" "$(grep ^osrelease /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "pot.potbase" "pot.potbase=test-pot" "$(grep ^pot.potbase /tmp/jails/new-pot/conf/pot.conf)"
-	assertEquals "ip4" "ip4=inherit" "$(grep ^ip4= /tmp/jails/new-pot/conf/pot.conf)"
+	assertEquals "ip" "" "$(grep ^ip= /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "vnet" "vnet=false" "$(grep ^vnet= /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "mkdir calls" "1" "$MKDIR_CALLS"
 	assertEquals "mkdir arg2" "${POT_FS_ROOT}/jails/new-pot/conf" "$MKDIR_CALL1_ARG2"
@@ -157,7 +160,7 @@ test_cj_conf_004()
 
 test_cj_conf_005()
 {
-	_cj_conf new-pot 11.1 inherit NO 2 inherit multi test-pot-2
+	_cj_conf new-pot 11.1 inherit inherit 2 inherit multi test-pot-2
 	assertEquals "return code" "0" "$?"
 	assertEquals "fscomp args1" "zpot/bases/11.1 /tmp/jails/new-pot/m ro" "$(sed '1!d' /tmp/jails/new-pot/conf/fscomp.conf)"
 	assertEquals "fscomp args2" "zpot/jails/test-pot/usr.local /tmp/jails/new-pot/m/usr/local ro" "$(sed '2!d' /tmp/jails/new-pot/conf/fscomp.conf)"
@@ -166,7 +169,8 @@ test_cj_conf_005()
 	assertEquals "pot.base" "pot.base=11.1" "$(grep ^pot.base /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "osrelease" "osrelease=\"11.1-RELEASE\"" "$(grep ^osrelease /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "pot.potbase" "pot.potbase=test-pot-2" "$(grep ^pot.potbase /tmp/jails/new-pot/conf/pot.conf)"
-	assertEquals "ip4" "ip4=inherit" "$(grep ^ip4= /tmp/jails/new-pot/conf/pot.conf)"
+	assertEquals "network_type" "network_type=inherit" "$(grep ^network_type= /tmp/jails/new-pot/conf/pot.conf)"
+	assertEquals "ip" "" "$(grep ^ip= /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "vnet" "vnet=false" "$(grep ^vnet= /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "mkdir calls" "1" "$MKDIR_CALLS"
 	assertEquals "mkdir arg2" "${POT_FS_ROOT}/jails/new-pot/conf" "$MKDIR_CALL1_ARG2"
@@ -180,7 +184,7 @@ test_cj_conf_005()
 
 test_cj_conf_006()
 {
-	_cj_conf new-pot 11.1 10.1.2.3 NO 1 inherit multi
+	_cj_conf new-pot 11.1 public-bridge 10.1.2.3 1 inherit multi
 	assertEquals "return code" "0" "$?"
 	assertEquals "fscomp args1" "zpot/bases/11.1 /tmp/jails/new-pot/m ro" "$(sed '1!d' /tmp/jails/new-pot/conf/fscomp.conf)"
 	assertEquals "fscomp args2" "zpot/jails/new-pot/usr.local /tmp/jails/new-pot/m/usr/local zfs-remount" "$(sed '2!d' /tmp/jails/new-pot/conf/fscomp.conf)"
@@ -189,7 +193,8 @@ test_cj_conf_006()
 	assertEquals "pot.base" "pot.base=11.1" "$(grep ^pot.base /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "osrelease" "osrelease=\"11.1-RELEASE\"" "$(grep ^osrelease /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "pot.potbase" "pot.potbase=" "$(grep ^pot.potbase /tmp/jails/new-pot/conf/pot.conf)"
-	assertEquals "ip4" "ip4=10.1.2.3" "$(grep ^ip4= /tmp/jails/new-pot/conf/pot.conf)"
+	assertEquals "network_type" "network_type=public-bridge" "$(grep ^network_type= /tmp/jails/new-pot/conf/pot.conf)"
+	assertEquals "ip" "ip=10.1.2.3" "$(grep ^ip= /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "vnet" "vnet=true" "$(grep ^vnet= /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "mkdir calls" "1" "$MKDIR_CALLS"
 	assertEquals "mkdir arg2" "${POT_FS_ROOT}/jails/new-pot/conf" "$MKDIR_CALL1_ARG2"
@@ -203,7 +208,7 @@ test_cj_conf_006()
 
 test_cj_conf_007()
 {
-	_cj_conf new-pot 11.1 inherit NO 1 pot multi
+	_cj_conf new-pot 11.1 inherit inherit 1 pot multi
 	assertEquals "return code" "0" "$?"
 	assertEquals "fscomp args1" "zpot/bases/11.1 /tmp/jails/new-pot/m ro" "$(sed '1!d' /tmp/jails/new-pot/conf/fscomp.conf)"
 	assertEquals "fscomp args2" "zpot/jails/new-pot/usr.local /tmp/jails/new-pot/m/usr/local zfs-remount" "$(sed '2!d' /tmp/jails/new-pot/conf/fscomp.conf)"
@@ -212,7 +217,8 @@ test_cj_conf_007()
 	assertEquals "pot.base" "pot.base=11.1" "$(grep ^pot.base /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "osrelease" "osrelease=\"11.1-RELEASE\"" "$(grep ^osrelease /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "pot.potbase" "pot.potbase=" "$(grep ^pot.potbase /tmp/jails/new-pot/conf/pot.conf)"
-	assertEquals "ip4" "ip4=inherit" "$(grep ^ip4= /tmp/jails/new-pot/conf/pot.conf)"
+	assertEquals "network_type" "network_type=inherit" "$(grep ^network_type= /tmp/jails/new-pot/conf/pot.conf)"
+	assertEquals "ip" "" "$(grep ^ip= /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "vnet" "vnet=false" "$(grep ^vnet= /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "pot.depend" "pot.depend=${POT_DNS_NAME}" "$(grep ^pot.depend /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "mkdir calls" "1" "$MKDIR_CALLS"
@@ -227,7 +233,7 @@ test_cj_conf_007()
 
 test_cj_conf_008()
 {
-	_cj_conf new-pot 11.1 10.1.2.3 NO 1 pot multi
+	_cj_conf new-pot 11.1 public-bridge 10.1.2.3 1 pot multi
 	assertEquals "return code" "0" "$?"
 	assertEquals "fscomp args1" "zpot/bases/11.1 /tmp/jails/new-pot/m ro" "$(sed '1!d' /tmp/jails/new-pot/conf/fscomp.conf)"
 	assertEquals "fscomp args2" "zpot/jails/new-pot/usr.local /tmp/jails/new-pot/m/usr/local zfs-remount" "$(sed '2!d' /tmp/jails/new-pot/conf/fscomp.conf)"
@@ -236,7 +242,8 @@ test_cj_conf_008()
 	assertEquals "pot.base" "pot.base=11.1" "$(grep ^pot.base /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "osrelease" "osrelease=\"11.1-RELEASE\"" "$(grep ^osrelease /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "pot.potbase" "pot.potbase=" "$(grep ^pot.potbase /tmp/jails/new-pot/conf/pot.conf)"
-	assertEquals "ip4" "ip4=10.1.2.3" "$(grep ^ip4= /tmp/jails/new-pot/conf/pot.conf)"
+	assertEquals "network_type" "network_type=public-bridge" "$(grep ^network_type= /tmp/jails/new-pot/conf/pot.conf)"
+	assertEquals "ip" "ip=10.1.2.3" "$(grep ^ip= /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "vnet" "vnet=true" "$(grep ^vnet= /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "pot.depend" "pot.depend=${POT_DNS_NAME}" "$(grep ^pot.depend /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "mkdir calls" "1" "$MKDIR_CALLS"
@@ -251,7 +258,7 @@ test_cj_conf_008()
 
 test_cj_conf_009()
 {
-	_cj_conf new-pot 11.1 10.1.2.3 YES 1 pot multi
+	_cj_conf new-pot 11.1 alias 10.1.2.3 1 pot multi
 	assertEquals "return code" "0" "$?"
 	assertEquals "fscomp args1" "zpot/bases/11.1 /tmp/jails/new-pot/m ro" "$(sed '1!d' /tmp/jails/new-pot/conf/fscomp.conf)"
 	assertEquals "fscomp args2" "zpot/jails/new-pot/usr.local /tmp/jails/new-pot/m/usr/local zfs-remount" "$(sed '2!d' /tmp/jails/new-pot/conf/fscomp.conf)"
@@ -260,7 +267,8 @@ test_cj_conf_009()
 	assertEquals "pot.base" "pot.base=11.1" "$(grep ^pot.base /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "osrelease" "osrelease=\"11.1-RELEASE\"" "$(grep ^osrelease /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "pot.potbase" "pot.potbase=" "$(grep ^pot.potbase /tmp/jails/new-pot/conf/pot.conf)"
-	assertEquals "ip4" "ip4=10.1.2.3" "$(grep ^ip4= /tmp/jails/new-pot/conf/pot.conf)"
+	assertEquals "network_type" "network_type=alias" "$(grep ^network_type= /tmp/jails/new-pot/conf/pot.conf)"
+	assertEquals "ip" "ip=10.1.2.3" "$(grep ^ip= /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "vnet" "vnet=false" "$(grep ^vnet= /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "pot.depend" "pot.depend=${POT_DNS_NAME}" "$(grep ^pot.depend /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "mkdir calls" "1" "$MKDIR_CALLS"
@@ -275,31 +283,7 @@ test_cj_conf_009()
 
 test_cj_conf_020()
 {
-	_cj_conf new-pot 11.1 inherit YES 1 pot multi
-	assertEquals "return code" "0" "$?"
-	assertEquals "fscomp args1" "zpot/bases/11.1 /tmp/jails/new-pot/m ro" "$(sed '1!d' /tmp/jails/new-pot/conf/fscomp.conf)"
-	assertEquals "fscomp args2" "zpot/jails/new-pot/usr.local /tmp/jails/new-pot/m/usr/local zfs-remount" "$(sed '2!d' /tmp/jails/new-pot/conf/fscomp.conf)"
-	assertEquals "fscomp args3" "zpot/jails/new-pot/custom /tmp/jails/new-pot/m/opt/custom zfs-remount" "$(sed '3!d' /tmp/jails/new-pot/conf/fscomp.conf)"
-	assertEquals "pot.level" "pot.level=1" "$(grep ^pot.level /tmp/jails/new-pot/conf/pot.conf)"
-	assertEquals "pot.base" "pot.base=11.1" "$(grep ^pot.base /tmp/jails/new-pot/conf/pot.conf)"
-	assertEquals "osrelease" "osrelease=\"11.1-RELEASE\"" "$(grep ^osrelease /tmp/jails/new-pot/conf/pot.conf)"
-	assertEquals "pot.potbase" "pot.potbase=" "$(grep ^pot.potbase /tmp/jails/new-pot/conf/pot.conf)"
-	assertEquals "ip4" "ip4=inherit" "$(grep ^ip4= /tmp/jails/new-pot/conf/pot.conf)"
-	assertEquals "vnet" "vnet=false" "$(grep ^vnet= /tmp/jails/new-pot/conf/pot.conf)"
-	assertEquals "pot.depend" "pot.depend=${POT_DNS_NAME}" "$(grep ^pot.depend /tmp/jails/new-pot/conf/pot.conf)"
-	assertEquals "mkdir calls" "1" "$MKDIR_CALLS"
-	assertEquals "mkdir arg2" "${POT_FS_ROOT}/jails/new-pot/conf" "$MKDIR_CALL1_ARG2"
-	assertEquals "internal_conf calls" "1" "$ICONF_CALLS"
-	assertEquals "internal_conf arg1" "new-pot" "$ICONF_CALL1_ARG1"
-	assertEquals "internal_conf arg2" "multi" "$ICONF_CALL1_ARG2"
-	assertEquals "internal_conf arg3" "1" "$ICONF_CALL1_ARG3"
-	assertEquals "internal_conf arg4" "inherit" "$ICONF_CALL1_ARG4"
-	assertEquals "sed calls" "0" "$SED_CALLS"
-}
-
-test_cj_conf_040()
-{
-	_cj_conf new-pot 11.1 inherit NO 0 pot single
+	_cj_conf new-pot 11.1 inherit inherit 0 pot single
 	assertEquals "return code" "0" "$?"
 	assertEquals "fscomp args1" "" "$(sed '1!d' /tmp/jails/new-pot/conf/fscomp.conf)"
 	assertEquals "fscomp args2" "" "$(sed '2!d' /tmp/jails/new-pot/conf/fscomp.conf)"
@@ -308,7 +292,8 @@ test_cj_conf_040()
 	assertEquals "pot.base" "pot.base=11.1" "$(grep ^pot.base /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "osrelease" "osrelease=\"11.1-RELEASE\"" "$(grep ^osrelease /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "pot.potbase" "pot.potbase=" "$(grep ^pot.potbase /tmp/jails/new-pot/conf/pot.conf)"
-	assertEquals "ip4" "ip4=inherit" "$(grep ^ip4= /tmp/jails/new-pot/conf/pot.conf)"
+	assertEquals "network_type" "network_type=inherit" "$(grep ^network_type= /tmp/jails/new-pot/conf/pot.conf)"
+	assertEquals "ip" "" "$(grep ^ip= /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "vnet" "vnet=false" "$(grep ^vnet= /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "pot.depend" "pot.depend=${POT_DNS_NAME}" "$(grep ^pot.depend /tmp/jails/new-pot/conf/pot.conf)"
 	assertEquals "mkdir calls" "1" "$MKDIR_CALLS"